support for overriding getactioncost function in child classes
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@12840 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 50afa037c171b1d27fc82540f38dab3048f6a1d0

+++ motion_planning/sbpl/src/test/test_adjacency_list.cpp
@@ -7,7 +7,7 @@ struct Point2D
 {
   Point2D(int newX, int newY) : x(newX), y(newY) {}
 
-  int heuristicDistanceTo (const Point2D& p) { int dx=p.x-x; int dy=p.y-y; int dist=(sqrt(dx*dx+dy*dy));  return dist; }
+  int heuristicDistanceTo (const Point2D& p) { int dx=p.x-x; int dy=p.y-y; int dist=((int)sqrt(dx*dx+dy*dy));  return dist; }
 
   int x;
   int y;
